1541: PrivateKeyCache: read key paths from envvars
For now we have chosen to leave the two implementations, PrivateKeyCache
and PrivateKeyCacheProvider in the codebase. The PrivateKeyCacheProvider
can be removed when it is no longer being used.

The biggest difference between these is the PrivateKeyCache is eager
about loading the the key files. Unlike the Guice provider pattern. This
simplifies the implementation and allows us to error if the keys are not
available (once keys are loaded they can safely be removed from disk).

The PrivateKeyCache will also load private keys from paths specified in
given environment variables.

Pair: DavidK / Archanaa