bau: add bare post() method to JsonClient
There is currently no way to run a POST request through JsonClient
without either appending a `sessionId` query parameter to the request,
or trying to capture the response into an object.  This results in fun
situations like in EventSinkHttpProxy [1] where we add a `sessionId`
query parameter to a service that doesn't use it at all.

We should fix this by adding a method which doesn't require either a
sessionid or a return Class<T> type.

[1] https://github.gds/gds/ida-hub/blob/release_134/shared/shared-rest/src/main/java/uk/gov/ida/shared/rest/event/sink/EventSinkHttpProxy.java#L44