refactor: remove sessionid functionality from JsonClient
See gds/ida-hub#237 for the argument as to why this was not useful.

This is still used in ida-frontend, but it shouldn't be too hard to
refactor it to use the non-sessionId methods.  ida-msa and ida-stub-idp
each have their own copy of this class, which should probably die at
some point.