1756: ability for JsonClient to return entity and cookies
When frontend is calling saml-proxy's SamlMessageReceiverApi, it needs to retrieve
the Session id entity and secure cookie created by saml-proxy. There is currently
no way to retrieve both so this commit attempts to rectify that. Ideally we would
like to use Scala's Tuple but none exists in Java so we've used a Map.Entry instead.