refactor: remove serviceName from DependentServiceConfiguration
We spend a fair amount of effort in puppet making sure every service
knows the serviceName of each of its dependent services.  We use this
information in the DependentServiceHealthCheck to ensure that not only
are we able to connect to the remote service, but also that it is the
one we are expecting.

My thoughts about this:

 * checking the serviceName of the remote service will catch a class of
   error (a misconfiguration) that the healthcheck wouldn't catch
   otherwise
 * on the other hand, smoke tests are also there to ensure our system is
   wired up correctly
 * we spend a lot of webops effort to wire up the serviceName to our
   apps (including making it configurable for apps which consume
   event-sink in case they consume stub-event-sink instead)
 * I'm not convinced the effort is worth the reward.

Therefore, I think we should just delete this check, and remove all the
code from ida-webops which sets serviceNames on services from
ida-webops.

If we do this, we need to also remember to do the equivalent change to
gds/doc-checking, which has a copy of DependentServiceHealthCheck.