refactor: fix serviceName backwards compatibility
In order for DropWizard to allow a configuration item, you seem to
need *both* a field *and* a getter for the field.  If you are missing
either one, you get an error message like the following:

 * Unrecognized field at: eventSink.serviceName

This re-adds the serviceName getter.  Also removes it from the builder
to emphasize it shouldn't be used any more.