bau: delete DependentServiceConfiguration and hangers-on
DependentServiceHealthCheck is no longer needed (after gds/ida-hub#319
is merged) and all this code can be removed. Hurrah! :fire:

getDependentServiceJerseyClientConfiguration() as a method can be
removed because it was only used by
DependentServiceHealthCheckClientProvider, removed in 48aa379de192a.

getClientTrustStoreConfiguration() can be removed from the
RestfulClientConfiguration interface because it was only needed by
DependentServiceHealthCheckBundle, removed in the same commit.