2033: Don't throw custom exceptions for NETWORK_ERROR
The only reason we can see that ever throw custom exceptions is to
ensure that they get "audited" which seems to be codespeak for "sent to
event-sink".  However, we specifically avoid sending NETWORK_ERROR
events to event-sink because of a problem, described in
gds/ida-hub@6d40537dc42, where failing to send an event to event-sink
would trigger another event to send, which would fail for the same
reason, which would trigger another event ad infinitum.

Without auditing, there seems to be no reason to wrap the
ClientHandlerException with a custom ApplicationException; it will
propagate to the top-level exception mapper and get logged as an error
anyway, so the ApplicationException adds nothing but extra noise to the
stacktrace.

Pair: @christopherholmes @philippotter