Revert "2033: Don't throw custom exceptions for NETWORK_ERROR"
This reverts commit c72b8035cfe6f82211988e56c16a3bfe1af4e0f0.

On reflection, there is one difference between throwing an
ApplicationException with type NETWORK_ERROR and throwing a
ClientHandlerException: an ApplicationException will be transformed into
an ErrorStatusDto with type NETWORK_ERROR, and communicate the problem
downstream to the consuming service.

A ClientHandlerException, on the other hand, will generally hit the
top-level ExceptionMapper.  In the case of SamlProxyExceptionMapper,
this will return an ErrorStatusDto with type UNKNOWN.  In the case of
Policy, there is *no* top-level ExceptionMapper<Exception> defined, so
this will probably just blow up.

We still think that this code should be removed, but it needs a little
more thought.  One option would be to create a
ClientHandlerExceptionMapper to handle the cases which were previously
handled by *ApplicationExceptionMapper with type NETWORK_ERROR.

Pair: @philippotter @christopherholmes

:'(