2363: Add http scheme to SecureSSLSchemeRegistryBuilder
We created an MSA truststore for saml-soap-proxy so it verifies the TLS cert from the MSA (see ida-hub commit
6147d13edb682cc4fab0a7471383d3951c8ea4cb). This truststore goes into the SecureSSLSchemeRegistryBuilder.

We tested the above in perf and found we couldn't get past the /response-processing page. This was because of a failure to communicate with the
config service. The error message was 'Caused by: com.sun.jersey.api.client.ClientHandlerException: java.lang.IllegalStateException: Scheme 'http' not registered.'

Therefore we have registered http scheme to the SecureSSLSchemeRegistryBuilder in order to allow communication to config.

Committing this to master because we need to test on perf; if this fix doesn't work it will be reverted.

pair: @adigoatabo @oswaldquek