2362: Allow client providers to specify a hostname verifier
BaseClientProvider is used by all the hub microservices; however we want to enable saml soap proxy to specifically verify
hostnames of MSAs being connected to match the CN in the MSA's TLS Cert. This can be done by using the
org.apache.http.conn.ssl.StrictHostnameVerifier. In order to allow saml soap proxy to use a different hostname verifier,
we're changing the constructor of BaseClientProvider to require a hostname verifier to be passed in.

pair: @adigoatabo @oswaldquek