2362: set the hostname verifier inside SSLSocketFactory
Previously an 'AllowAllHostnameVerifier' was set as part of an instantiation of IdaJerseyClientBuilder. This happened
inside BaseClientProvider. However when the 'AllowAllHostnameVerifier' was changed to 'StrictHostnameVerifier'
connections from saml-soap-proxy to MSA on perf didn't fail as expected. This led us to believe setting a hostname
verifier inside IdaJerseyClientBuilder didn't have any effect. This was proved right when it was set inside the
SSLSocketFactory that is instantiated inside SecureSSLSchemeRegistryBuilder.