bau: Add filter to clear MDC after every request
The [MDC][] (or Mapped Diagnostic Context) is essentially a thread-local
mutable HashMap where the application can store data and the logging
framework can retrieve it.  gds/ida-msa#47 provides an example of using
the MDC to log a SAML request ID whenever an error is encountered in
processing the SAML message.

Because the MDC is mutable and thread-local, and because our application
reuses the same threads for multiple requests, data in the MDC might
hang around from one request to the next on the same thread.  It is
therefore important for us to clean up the MDC after ourselves.

For data which can easily be acquired from the http request (such as
headers), you can ensure you add and remove data from the MDC in the
same place, in a servlet filter.  An example of this is in DCS's
[RequestContextFilter][].

However, for SAML request IDs, we can't use the same trick, because the
SAML request is in the request body which shouldn't really be processed
by a servlet filter.  This means that the SAML request ID needs to be
added to the MDC lower down the stack, at the point where the SAML
request gets parsed ([example][request-mapper-example]), and then
retrieved higher up the stack, in the ExceptionMapper which is
responsible for logging the particular exception.

This means we can't do the MDC.put() / try / finally / MDC.remove()
pattern of DCS's RequestContextFilter.

As a rather blunt method of cleaning up the MDC in this case, this
commit adds a ClearMdcAfterRequestFilter class which just wipes the
entire MDC clean after every request.  (It does so *after*, rather
than *before*, the request to avoid issues with filter ordering.  If one
filter adds an item to the MDC before a request, and another clears the
whole MDC, it becomes very important which order the filters are added.
Clearing the MDC after the request makes this problem go away.)

[MDC]: http://logback.qos.ch/manual/mdc.html
[RequestContextFilter]: https://github.gds/gds/doc-checking/blob/1fcef37f0b99caae7f34ffb92148ab67091c2687/shared/src/main/java/uk/gov/ida/docchecking/shared/context/RequestContextFilter.java
[request-mapper-example]: https://github.gds/gds/ida-msa/blob/5a9a49a8172a15441bacee9f013787ccf8db108e/src/main/java/uk/gov/ida/matchingserviceadapter/mappers/DocumentToInboundMatchingServiceRequestMapper.java#L42