bau: Include uri in error message
JsonResponseProcessor is the only place we create an `ApplicationException` from
the `ErrorStatusDto`. (Currently there is a copy of this in `ida-msa` but that will
soon be removed!). Here we always have the uri we are trying to contact so we should log
that as part of the error message. This should make error messages more useful
as we will be able to see the url we are trying to contact when something goes
  wrong.