2159: remove Analytics dispatcher stuff
Jersey did not port over the dispatcher adapter features from version 1 to 2 so removing all this.
It is not worth the effort to attempt to recreate the feature for ourselves as it has minimal impact on frontend.
Minimal impact meaning that frontend will need to make only 5 manual calls to AnalyticsReporter which is not a lot of effort at all.