3652: Make the PiwikClient's Client Named
The Client is used by our services to talk to each other. It is being updated
to use a truststore. This updated Client will no longer work for communications
between frontend and piwik. We have therefore put a Named annotation on the
PiwikClient's Client so dependant projects will have to provide a custom client
to PiwikClient.

@oswaldquek @pauloschneider