5466 Add retry functionality to ErrorHandlingClient
ErrorHandlingClient can now be passed an integer which configures the
number of times to retry the http request.

At the moment this functionality is only configurable at constructor
time. It may be prefeable to configure this value on a per requst basis:
e.g. a single instance of the Client may want seperate values for each
request it makes.